refactor(game): refactoring project to modular strucuture
