+++ src/game/__init__.py
@@ -1,0 +1,6 @@
+from .routes import router
+
+
+__all__ = [
+    "router"
+]
