+++ src/game/routes.py
@@ -4,10 +4,21 @@ from fastapi import Body
 
 from src.database.connection import DbConnectionHandler
 from src.api import dependencies
-from src.domain import game
-from src import schemas
+
+from . import domain
+from . import schemas
+
 
 router = APIRouter()
+
+
+@router.get("/", response_model=list[schemas.Game])
+async def list_game(
+    context: DbConnectionHandler = Depends(
+        dependencies.get_database_connection)
+):
+    return await domain.ListGameUseCase(context).execute()
+
 
 
 @router.post("/game", response_model=schemas.Game)
@@ -15,4 +26,4 @@ async def crete_game(
     context: DbConnectionHandler = Depends(dependencies.get_database_connection),
     payload: schemas.Game = Body(...)
 ):
-    return await game.CreateGameUseCase(context, payload).execute()
+    return await domain.CreateGameUseCase(context, payload).execute()
