--- src/models/__init__.py
@@ -1,6 +1,0 @@
-from .game import GameModel
-
-
-__all__ = [
-    "GameModel"
-]
