--- src/routes/__init__.py
@@ -1,6 +1,0 @@
-from .core import router
-
-
-__all__ = [
-    "router"
-]
