--- src/schemas/__init__.py
@@ -1,8 +1,0 @@
-from .game import Game
-from .game import CreateGame
-
-
-__all__ = [
-    "Game",
-    "CreateGame" 
-]
